requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-js-asset==1.2.2
django-storages==1.9.1
dropbox==9.4.0
gunicorn==20.0.4
idna==2.9
Pillow==7.0.0
psycopg2==2.8.4
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
